refactor: Split `AccountSwitch` in multiple modules
