+++ src/ducks/account/AccountSwitchWrapper.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import cx from 'classnames'
+
+import styles from 'ducks/account/AccountSwitch.styl'
+
+const AccountSwitchWrapper = ({ children }) => {
+  return <div className={cx(styles['account-switch'])}>{children}</div>
+}
+
+export default AccountSwitchWrapper
